Offers: split offer tests to two files
- Tests for searching offers
- Tests for reading/creating/updating offers
